Move JSON conversion code to detail
